requirements.txt
altair==4.2.0
install_jdk==0.3.0
pandas==1.4.2
streamlit==1.9.0
